pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Michael-McD <michael.mcdowell@hotmail.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.0"
fastapi = {extras = ["all"], version = "^0.104.1"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
